requirements.txt
fastapi[standard]
google-genai
sqlalchemy
python-dotenv
fastapi_sso.sso
sqlalchemy
